cleanup: removed all unused code that related to auth && fixed clippy errors
