+++ server/src/data/mod.rs
@@ -1,3 +1,2 @@
 pub mod models;
 pub mod schema;
-pub mod validators;
