--- server/src/data/validators.rs
@@ -1,4 +1,0 @@
-pub fn email_regex() -> regex::Regex {
-    regex::Regex::new(r"^[a-zA-Z0-9.!#$%&’*+/=?^_`{|}~-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*")
-        .unwrap()
-}
