+++ server/src/handlers/mod.rs
@@ -3,11 +3,8 @@ pub mod card_handler;
 pub mod collection_handler;
 pub mod dataset_handler;
 pub mod file_handler;
-pub mod invitation_handler;
 pub mod message_handler;
 pub mod notification_handler;
 pub mod organization_handler;
-pub mod password_reset_handler;
-pub mod register_handler;
 pub mod topic_handler;
 pub mod user_handler;
