+++ server/src/operators/email_operator.rs
@@ -1,65 +1,5 @@
-use crate::{data::models::PasswordReset, errors::DefaultError, get_env};
+use crate::{errors::DefaultError, get_env};
 use sendgrid::v3::{Content, Email, Message, Personalization, Sender};
-
-// pub fn send_invitation(app_url: String, invitation: &Invitation) -> Result<(), DefaultError> {
-//     let sg_email_content = format!(
-//         "Please click on the link below to complete registration. <br/>
-//          <a href=\"{}/auth/register/{}?email={}\">
-//          {}/register</a> <br>
-//          your Invitation expires at <strong>{}</strong>",
-//         app_url,
-//         invitation.id,
-//         invitation.email,
-//         app_url,
-//         invitation.expires_at.format("%I:%M %p %A, %-d %B, %C%y")
-//     );
-//     let sg_email_personalization = Personalization::new(Email::new(invitation.email.as_str()));
-//     let email_address = Email::new(get_env!("SENDGRID_EMAIL_ADDRESS", "SENDGRID_EMAIL_ADDRESS should be set"));
-//     let sg_email = Message::new(email_address)
-//         .set_subject("You have been invited to join Arguflow AI")
-//         .add_content(
-//             Content::new()
-//                 .set_content_type("text/html")
-//                 .set_value(sg_email_content),
-//         )
-//         .add_personalization(sg_email_personalization);
-
-//     send_email(sg_email)
-// }
-
-pub fn send_password_reset(
-    app_url: String,
-    password_reset: &PasswordReset,
-) -> Result<(), DefaultError> {
-    let sg_email_content = format!(
-        "Please click on the link below to reset your password. <br/>
-         <a href=\"{}/auth/password/{}?email={}\">
-         {}/auth/password</a> <br>
-         your password reset link expires at <strong>{}</strong>",
-        app_url,
-        password_reset.id,
-        password_reset.email,
-        app_url,
-        password_reset
-            .expires_at
-            .format("%I:%M %p %A, %-d %B, %C%y")
-    );
-    let sg_email_personalization = Personalization::new(Email::new(password_reset.email.as_str()));
-    let email_address = Email::new(get_env!(
-        "SENDGRID_EMAIL_ADDRESS",
-        "SENDGRID_EMAIL_ADDRESS should be set"
-    ));
-    let sg_email = Message::new(email_address)
-        .set_subject("Reset your Arguflow AI password")
-        .add_content(
-            Content::new()
-                .set_content_type("text/html")
-                .set_value(sg_email_content),
-        )
-        .add_personalization(sg_email_personalization);
-
-    send_email(sg_email)
-}
 
 #[allow(dead_code)]
 pub fn send_health_check_error(email: String, error: String) -> Result<(), DefaultError> {
