+++ server/src/operators/mod.rs
@@ -6,7 +6,6 @@ pub mod file_operator;
 pub mod message_operator;
 pub mod notification_operator;
 pub mod organization_operator;
-pub mod password_reset_operator;
 pub mod qdrant_operator;
 pub mod search_operator;
 pub mod tantivy_operator;
